fixed bg p-matrix writes not taking into account sign bit due to improper casting

+++ src/emu/source/hw/ppu/background.d
@@ -33,7 +33,7 @@ struct Background {
 
     Layer layer;
     
-    int[4] p;
+    short[4] p;
 }
 
 enum BackgroundMode {
